requirements.txt
pydantic
docx2txt
PyPDF2
sentence-transformers
transformers
faiss-cpu
numpy
protobuf
tiktoken